+++ cmd/jujud/agent/engine_test.go
@@ -105,6 +105,7 @@ var (
 		"agent",
 		"api-caller",
 		"api-config-watcher",
+		"clock",
 		"migration-fortress",
 		"migration-inactive-flag",
 		"migration-minion",
