+++ cmd/jujud/agent/internal/mongomaster/mock_test.go
@@ -1,7 +1,7 @@
 // Copyright 2017 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package master_test
+package mongomaster_test
 
 import "github.com/juju/testing"
